Delete the orchestrator and jax engine code.
